+++ pkg/storage/chunk/cache/groupcache.go
@@ -3,7 +3,6 @@ package cache
 import (
 	"context"
 	"crypto/tls"
-	"flag"
 	"fmt"
 	"net"
 	"net/http"
@@ -62,26 +61,13 @@ type RingCfg struct {
 type GroupCacheConfig struct {
 	Enabled           bool          `yaml:"enabled,omitempty"`
 	Ring              RingCfg       `yaml:"ring,omitempty"`
-	CapacityMB        int64         `yaml:"capacity_per_cache_mb,omitempty"`
+	MaxSizeMB         int64         `yaml:"max_size_mb,omitempty"`
 	ListenPort        int           `yaml:"listen_port,omitempty"`
 	HeartbeatInterval time.Duration `yaml:"heartbeat_interval,omitempty"`
 	HeartbeatTimeout  time.Duration `yaml:"heartbeat_timeout,omitempty"`
 	WriteByteTimeout  time.Duration `yaml:"write_timeout,omitempty"`
 
 	Cache Cache `yaml:"-"`
-}
-
-// RegisterFlagsWithPrefix adds the flags required to config this to the given FlagSet
-func (cfg *GroupCacheConfig) RegisterFlagsWithPrefix(prefix, _ string, f *flag.FlagSet) {
-	cfg.Ring.RegisterFlagsWithPrefix(prefix, "", f)
-
-	f.BoolVar(&cfg.Enabled, prefix+".enabled", false, "Whether or not groupcache is enabled")
-	f.IntVar(&cfg.ListenPort, prefix+".listen_port", 4100, "The port to use for groupcache communication")
-	f.Int64Var(&cfg.CapacityMB, prefix+".capacity-per-cache-mb", 100, "Capacity of each groupcache group in MB (default: 100). "+
-		"NOTE: there are 3 caches (result, chunk, and index query), so the maximum used memory will be *triple* the value specified here.")
-	f.DurationVar(&cfg.HeartbeatInterval, "prefix.heartbeat-interval", time.Second, "If the connection is idle, the interval the cache will send heartbeats")
-	f.DurationVar(&cfg.HeartbeatTimeout, "prefix.heartbeat-timeout", time.Second, "Timeout for heartbeat responses")
-	f.DurationVar(&cfg.WriteByteTimeout, "prefix.write-timeout", time.Second, "Maximum time for the cache to try writing")
 }
 
 type ringManager interface {
@@ -116,7 +102,7 @@ func NewGroupCache(rm ringManager, config GroupCacheConfig, logger log.Logger, r
 		wg:                   sync.WaitGroup{},
 		startWaitingForClose: cancel,
 		reg:                  reg,
-		cacheBytes:           config.CapacityMB * 1e6, // MB => B
+		cacheBytes:           config.MaxSizeMB * 1e6, // MB => B
 	}
 
 	go cache.serveGroupcache(config.ListenPort)
@@ -208,7 +194,7 @@ func (c *GroupCache) Stats() *groupcache.Stats {
 
 // Groupconfig represents config per Group.
 type GroupConfig struct {
-	CapacityMB int64 `yaml:"capacity_mb,omitempty"`
+	MaxSizeMB int64 `yaml:"max_size_mb,omitempty"`
 }
 
 type group struct {
@@ -234,8 +220,8 @@ func (c *GroupCache) NewGroup(name string, cfg *GroupConfig, ct stats.CacheType)
 	c.startWaitingForClose()
 
 	cap := c.cacheBytes
-	if cfg.CapacityMB != 0 {
-		cap = cfg.CapacityMB * 1e6 // MB into bytes
+	if cfg.MaxSizeMB != 0 {
+		cap = cfg.MaxSizeMB * 1e6 // MB into bytes
 	}
 
 	requestDuration := promauto.With(c.reg).NewHistogramVec(prometheus.HistogramOpts{