+++ tools/lambda-promtail/main.tf
@@ -66,6 +66,7 @@ resource "aws_lambda_function" "lambda_promtail" {
   image_uri     = var.lambda_promtail_image
   function_name = "lambda_promtail"
   role          = aws_iam_role.iam_for_lambda.arn
+  kms_key_arn   = var.kms_key_arn
 
   timeout      = 60
   memory_size  = 128