+++ operator/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Main
 
+- [6999](https://github.com/grafana/loki/pull/6999) **Red-GV**: Adding LokiStack Gateway alerts
+- [7000](https://github.com/grafana/loki/pull/7000) **xperimental**: Configure default node affinity for all pods
+- [6923](https://github.com/grafana/loki/pull/6923) **xperimental**: Reconcile owner reference for existing objects
+- [6907](https://github.com/grafana/loki/pull/6907) **Red-GV**: Adding valid subscription annotation to operator metadata
 - [6479](https://github.com/grafana/loki/pull/6749) **periklis**: Update Loki operand to v2.6.1
 - [6748](https://github.com/grafana/loki/pull/6748) **periklis**: Update go4.org/unsafe/assume-no-moving-gc to latest
 - [6741](https://github.com/grafana/loki/pull/6741) **aminesnow**: Golang version to 1.18 and k8s client to 1.24